Start BlockSectionValidationSpecification
